Instruction regarding development on Windows
Added instruction on how to develop the package on Windows as opposed to
Linux, with notes that I would have needed when I tried to do it.

This should help users in the long run because whilst the package is
set to be depreciated in the future development is still required
because of how closely it is tied to pymapdl.